pyproject.toml
[tool.poetry]
name = "pynamics365"
version = "0.1.0"
description = ""
authors = ["Dan Lawson <52593003+danlsn@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
requests-cache = "^1.0.1"
tqdm = "^4.65.0"
pandas = "^1.5.3"
aiofiles = "^23.1.0"
httpx = "^0.23.3"
aiopath = "^0.6.11"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
